Add gpio.(*Bank).SetHold(g int) support.
There are situations where you need to atomically set a value of
a GPIO from the read values of some other GPIOs. To support
concurrent GPIO access, such an operation requires us to hold
the write lock for the duration of such a computation.

Integrate upstream fix for not noticing x -> 0 transitions in
output iotracer VCD dumps.

Signed-off-by: Tinkerer <tinkerer@zappem.net>